Return length of collected arrays & linked lists
Adds a parameter to both citer_collect_into_array and
citer_collect_into_linked_list.

For the array collector, this parameter is a pointer to a size_t
variable in which the length of the returned array is stored.

For the linked list collector, this parameter is a pointer to a pointer
to a citer_llnode_t, in which the pointer to the tail of the list is
stored.