Start on fuzzer: intercept malloc(3) and free(3)
Since the fuzzer will randomly call citer_*() functions, we can't track
all dynamically-allocated memory. Since the point of the fuzzer is to
run for a long time, any memory leaks could cause problems. Thus we need
a way to garbage-collect.

This commit creates a framework which intercepts all malloc(3) and
free(3) calls to give us a list of items which are still allocated. A
free_all() function is also provided, which will free all outstanding
allocations.